+++ drizzle.config.js
@@ -9,6 +9,6 @@ module.exports = defineConfig({
   out: './drizzle',
   dialect: 'postgresql',
   dbCredentials: {
-    url: process.env.DATABASE_URL,
+    url: process.env.POSTGRES_URL || process.env.DATABASE_URL,
   },
 });